refactor: このJWTはSalesのStaffをコンテキストとするので、Sales配下に移動
