+++ apps/nest-app/src/sale/ui/auth/auth.service.spec.ts
