+++ apps/nest-app/src/sale/ui/auth/auth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { Staff, UsersService } from '../users/users.service';
+import { Staff, UsersService } from '../../../users/users.service';
 import { JwtService } from '@nestjs/jwt';
 
 @Injectable()
