+++ apps/nest-app/src/sale/ui/auth/constants.ts
