+++ apps/nest-app/src/sale/ui/auth/jwt.strategy.ts
